test: move the ui tests from a flow package to feature
Move into a features package which better reflects what each 'feature' is
